src/requirements.txt
flask>=3.0,<3.2
gunicorn>=22,<24
python-magic>=0.4.27,<0.5